+++ app/src/test/java/com/storehop/app/data/repository/CategoryRepositoryImplTest.kt
@@ -2,8 +2,10 @@ package com.storehop.app.data.repository
 
 import com.google.common.truth.Truth.assertThat
 import com.storehop.app.data.db.StorehopDatabase
+import com.storehop.app.analytics.AnalyticsService
 import com.storehop.app.data.util.FakeHouseholdSessionProvider
 import com.storehop.app.data.util.IdGenerator
+import io.mockk.mockk
 import com.storehop.app.testing.FakeSessionProvider
 import com.storehop.app.testing.TEST_USER_ID
 import com.storehop.app.testing.createTestDb
@@ -38,6 +40,7 @@ class CategoryRepositoryImplTest {
             clock = Clock.fixed(Instant.ofEpochMilli(50_000L), ZoneOffset.UTC),
             session = FakeSessionProvider("local-only"),
             householdSession = FakeHouseholdSessionProvider("local-only"),
+            analytics = mockk(relaxed = true),
         )
     }
     @After fun tearDown() { db.close() }
@@ -215,6 +218,7 @@ class CategoryRepositoryImplTest {
             clock = Clock.fixed(Instant.ofEpochMilli(50_000L), ZoneOffset.UTC),
             session = FakeSessionProvider("some-other-user"),
             householdSession = FakeHouseholdSessionProvider("some-other-user"),
+            analytics = mockk(relaxed = true),
         )
 
         otherRepo.rename("cat_produce", "HIJACKED")
@@ -287,6 +291,7 @@ class CategoryRepositoryImplTest {
             clock = Clock.fixed(Instant.ofEpochMilli(50_000L), ZoneOffset.UTC),
             session = FakeSessionProvider("some-other-user"),
             householdSession = FakeHouseholdSessionProvider("some-other-user"),
+            analytics = mockk(relaxed = true),
         )
 
         otherRepo.setArchived("cat_produce", archived = true)